tests/wpt/web-platform-tests/navigation-timing/nav2_test_navigation_type_navigate.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>Navigation Timing 2 WPT</title>
        <link rel="author" title="Google" href="http://www.google.com/" />
        <link rel="help" href="http://www.w3.org/TR/navigation-timing/#sec-navigation-timing-interface"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script>
            async_test(function (t) {
                var observer = new PerformanceObserver(
                    t.step_func(function (entryList) {
                        assert_equals(entryList.getEntries()[0].type, "navigate", "Expected navigation type to be navigate.");
                        observer.disconnect();
                        t.done();
                    })
                );
                observer.observe({entryTypes: ["navigation"]});

            }, "Navigation type to be navigate.");
        </script>

    </head>
    <body>
        <h1>
            Description</h1>
        <p>
            This test validates the value of window.performance.getEntriesByType("navigation")[0].type to be navigate.</p>

    </body>
</html>
